+++ src/redisPubSub.js
@@ -112,7 +112,7 @@ const broadcastEvent = async (tenantId, type, subType) => {
       [type, subType].join("##")
     );
     if (result) {
-      logger.info("skip publish redis event as no lock is available", {
+      logger.debug("skip publish redis event as no lock is available", {
         type,
         subType,
       });