+++ src/shared/WorkerQueue.js
@@ -4,14 +4,29 @@ const cds = require("@sap/cds");
 
 const config = require("../config");
 const EventQueueError = require("../EventQueueError");
+const { Priorities } = require("../constants");
+const SetIntervalDriftSafe = require("./SetIntervalDriftSafe");
+
+const PRIORITIES = Object.values(Priorities).reverse();
+const PRIORITY_MULTIPLICATOR = PRIORITIES.reduce((result, element, index) => {
+  result[element] = index + 1;
+  return result;
+}, {});
 
 const COMPONENT_NAME = "/eventQueue/WorkerQueue";
 const NANO_TO_MS = 1e6;
+const MIN_TO_MS = 60 * 1000;
+const INCREASE_PRIORITY_AFTER = 3;
+
+let lastLogTs;
+
 const THRESHOLD = {
   INFO: 35 * 1000,
   WARN: 55 * 1000,
   ERROR: 75 * 1000,
 };
+
+const CHECK_INTERVAL_QUEUE = 60 * 1000;
 
 class WorkerQueue {
   #concurrencyLimit;
@@ -28,24 +43,53 @@ class WorkerQueue {
     }
     this.#runningPromises = [];
     this.#runningLoad = 0;
-    this.#queue = [];
+    this.#queue = PRIORITIES.reduce((result, priority) => {
+      result[priority] = [];
+      return result;
+    }, {});
+
+    const runner = new SetIntervalDriftSafe(CHECK_INTERVAL_QUEUE);
+    runner.run(this.#adjustPriority.bind(this));
   }
 
-  addToQueue(load, label, cb) {
+  addToQueue(load, label, priority = Priorities.Medium, cb) {
     if (load > this.#concurrencyLimit) {
       throw EventQueueError.loadHigherThanLimit(load, label);
     }
 
+    if (!PRIORITIES.includes(priority)) {
+      throw EventQueueError.priorityNotAllowed(priority, label);
+    }
+
     const startTime = process.hrtime.bigint();
     const p = new Promise((resolve, reject) => {
-      this.#queue.push([load, label, cb, resolve, reject, startTime]);
+      this.#queue[priority].push([load, label, cb, resolve, reject, startTime]);
     });
-    this._checkForNext();
+    this.#checkForNext();
     return p;
   }
 
-  _executeFunction(load, label, cb, resolve, reject, startTime) {
-    this.checkAndLogWaitingTime(startTime, label);
+  #adjustPriority() {
+    const checkTime = process.hrtime.bigint();
+    const priorityValues = Object.values(Priorities);
+
+    for (let i = 0; i < priorityValues.length - 1; i++) {
+      const priority = priorityValues[i];
+      const nextPriority = priorityValues[i + 1];
+      for (let i = 0; i < this.queue[priority].length; i++) {
+        const queueEntry = this.queue[priority][i];
+        const startTime = queueEntry[6] ?? queueEntry[5];
+        if (Math.round(Number(checkTime - startTime) / NANO_TO_MS) > INCREASE_PRIORITY_AFTER * MIN_TO_MS) {
+          const [entry] = this.queue[priority].splice(i, 1);
+          entry.push(checkTime);
+          this.queue[nextPriority].push(entry);
+        }
+      }
+    }
+  }
+
+  _executeFunction(load, label, cb, resolve, reject, startTime, priority) {
+    this.#checkAndLogWaitingTime(startTime, label, priority);
     const promise = Promise.resolve().then(() => cb());
     this.#runningPromises.push(promise);
     this.#runningLoad = this.#runningLoad + load;
@@ -53,7 +97,7 @@ class WorkerQueue {
       .finally(() => {
         this.#runningLoad = this.#runningLoad - load;
         this.#runningPromises.splice(this.#runningPromises.indexOf(promise), 1);
-        this._checkForNext();
+        this.#checkForNext();
       })
       .then((...results) => {
         resolve(...results);
@@ -62,15 +106,32 @@ class WorkerQueue {
         cds.log(COMPONENT_NAME).error("Error happened in WorkQueue. Errors should be caught before!", err, { label });
         reject(err);
       });
+
+    if (this.#runningLoad !== this.#concurrencyLimit) {
+      this.#checkForNext();
+    }
   }
 
-  _checkForNext() {
-    const load = this.#queue[0]?.[0];
-    if (!this.#queue.length || this.#runningLoad + load > this.#concurrencyLimit) {
+  #checkForNext() {
+    if (!this.#queue.length && this.#runningLoad === this.#concurrencyLimit) {
       return;
     }
-    const args = this.#queue.shift();
-    this._executeFunction(...args);
+
+    let entryFound = false;
+    for (const priority of PRIORITIES) {
+      for (let i = 0; i < this.#queue[priority].length; i++) {
+        const [load] = this.#queue[priority][i];
+        if (this.#runningLoad + load <= this.#concurrencyLimit) {
+          const [args] = this.#queue[priority].splice(i, 1);
+          this._executeFunction(...args, priority);
+          entryFound = true;
+          break;
+        }
+      }
+      if (entryFound) {
+        break;
+      }
+    }
   }
 
   get runningPromises() {
@@ -87,14 +148,24 @@ class WorkerQueue {
     return WorkerQueue.#instance;
   }
 
-  checkAndLogWaitingTime(startTime, label) {
+  get queue() {
+    return this.#queue;
+  }
+
+  #checkAndLogWaitingTime(startTime, label, priority) {
+    const ts = Date.now();
+    if (ts - lastLogTs <= 1000) {
+      return;
+    }
+    lastLogTs = ts;
     const diffMs = Math.round(Number(process.hrtime.bigint() - startTime) / NANO_TO_MS);
+    const priorityMultiplication = PRIORITY_MULTIPLICATOR[priority];
     let logLevel;
-    if (diffMs >= THRESHOLD.ERROR) {
+    if (diffMs >= THRESHOLD.ERROR * priorityMultiplication) {
       logLevel = "error";
-    } else if (diffMs >= THRESHOLD.WARN) {
+    } else if (diffMs >= THRESHOLD.WARN * priorityMultiplication) {
       logLevel = "warn";
-    } else if (diffMs >= THRESHOLD.INFO) {
+    } else if (diffMs >= THRESHOLD.INFO * priorityMultiplication) {
       logLevel = "info";
     } else {
       logLevel = "debug";