+++ test-integration/runner.test.js
@@ -25,6 +25,7 @@ const distributedLock = require("../src/shared/distributedLock");
 const eventQueue = require("../src");
 const runner = require("../src/runner");
 const path = require("path");
+const periodicEvents = require("../src/periodicEvents");
 
 const tenantIds = [
   "cd805323-879c-4bf7-b19c-8ffbbee22e1f",
@@ -78,13 +79,11 @@ describe("redisRunner", () => {
     getAllTenantIdsSpy
       .mockResolvedValueOnce(tenantIds)
       .mockResolvedValueOnce(tenantIds)
-      .mockResolvedValueOnce(tenantIds)
       .mockResolvedValueOnce(tenantIds);
     const p1 = runner.__._multiTenancyRedis();
     const p2 = runner.__._multiTenancyRedis();
 
-    const [promises1, promises2] = await Promise.allSettled([p1, p2]);
-    await Promise.allSettled(promises1.value.concat(promises2.value));
+    await Promise.allSettled([p1, p2]);
     await Promise.allSettled(WorkerQueue.instance.runningPromises);
 
     expect(setValueWithExpireSpy).toHaveBeenCalledTimes(3);
@@ -118,8 +117,7 @@ describe("redisRunner", () => {
     expect(tenantChecks).toMatchSnapshot();
 
     // another run within 5 minutes should do nothing
-    const promises = await runner.__._multiTenancyRedis();
-    await Promise.allSettled(promises);
+    await runner.__._multiTenancyRedis();
     await Promise.allSettled(WorkerQueue.instance.runningPromises);
     expect(acquireLockSpy).toHaveBeenCalledTimes(21);
     expect(processEventQueueSpy).toHaveBeenCalledTimes(6);
@@ -130,11 +128,12 @@ describe("redisRunner", () => {
   it("db", async () => {
     configInstance.redisEnabled = false;
     const originalCdsTx = cds.tx;
+    jest.spyOn(periodicEvents, "checkAndInsertPeriodicEvents").mockResolvedValue();
     jest.spyOn(cds, "tx").mockImplementation(function (context, fn) {
       if (!fn) {
         return originalCdsTx.call(this, context);
       }
-      if (!fn.toString().toLowerCase().includes("worker")) {
+      if (fn.toString().toLowerCase().includes("await fn(tx, ...parameters)")) {
         context.tenant = null;
       }
       return originalCdsTx.call(this, context, fn);
@@ -144,19 +143,17 @@ describe("redisRunner", () => {
       .mockResolvedValueOnce(tenantIds)
       .mockResolvedValueOnce(tenantIds)
       .mockResolvedValueOnce(tenantIds)
-      .mockResolvedValueOnce(tenantIds)
       .mockResolvedValueOnce(tenantIds);
     expect(processEventQueueSpy).toHaveBeenCalledTimes(0);
     const p1 = runner.__._multiTenancyDb();
     const p2 = runner.__._multiTenancyDb();
-    const [promises1, promises2] = await Promise.allSettled([p1, p2]);
-    await Promise.allSettled(promises1.value.concat(promises2.value));
+    await Promise.allSettled([p1, p2]);
     await Promise.allSettled(WorkerQueue.instance.runningPromises);
 
     // 3 tenants * 1 acquire lock for periodic run (happens only once per instance [tenant hash] +
     // 3 tenants * 2 fn calls * 2 events (1 ad hoc and 1 periodic)
     expect(acquireLockSpy).toHaveBeenCalledTimes(15);
-    // 3 tenants * 1 ad hoc and 1 periodic
+    // 3 tenants * 1 ad hoc + 1 periodic
     expect(processEventQueueSpy).toHaveBeenCalledTimes(6);
 
     const acquireLockMock = acquireLockSpy.mock;
@@ -183,8 +180,7 @@ describe("redisRunner", () => {
     expect(tenantChecks).toMatchSnapshot();
 
     // another run within 5 minutes should do nothing
-    const promises = await runner.__._multiTenancyDb();
-    await Promise.allSettled(promises);
+    await runner.__._multiTenancyDb();
     await Promise.allSettled(WorkerQueue.instance.runningPromises);
     expect(processEventQueueSpy).toHaveBeenCalledTimes(6);
     expect(acquireLockSpy).toHaveBeenCalledTimes(21);
@@ -198,8 +194,7 @@ describe("redisRunner", () => {
       )
     );
 
-    const promises3 = await runner.__._multiTenancyDb();
-    await Promise.allSettled(promises3);
+    await runner.__._multiTenancyDb();
     await Promise.allSettled(WorkerQueue.instance.runningPromises);
     expect(WorkerQueue.instance.runningPromises).toHaveLength(0);
     expect(acquireLockSpy).toHaveBeenCalledTimes(27);
@@ -215,7 +210,7 @@ describe("redisRunner", () => {
       if (!fn) {
         return originalCdsTx.call(this, context);
       }
-      if (!fn.toString().toLowerCase().includes("worker")) {
+      if (fn.toString().toLowerCase().includes("await fn(tx, ...parameters)")) {
         context.tenant = null;
       }
       return originalCdsTx.call(this, context, fn);
@@ -303,9 +298,8 @@ describe("redisRunner", () => {
           }
         });
       });
-      getAllTenantIdsSpy.mockResolvedValueOnce(tenantIds).mockResolvedValueOnce(tenantIds);
-      const p1 = runner.__._multiTenancyDb();
-      await p1.then((p) => Promise.allSettled(p));
+      getAllTenantIdsSpy.mockResolvedValueOnce(tenantIds);
+      await runner.__._multiTenancyDb();
       await promise;
       expect(acquireLockSpy.mock.calls.filter(([, key]) => key === "EVENT_QUEUE_RUN_PERIODIC_EVENT")).toHaveLength(3);
 
@@ -325,16 +319,14 @@ describe("redisRunner", () => {
           }
         });
       });
-      getAllTenantIdsSpy.mockResolvedValueOnce(tenantIds).mockResolvedValueOnce(tenantIds);
-      const p1 = runner.__._multiTenancyDb();
-      await p1.then((p) => Promise.allSettled(p));
+      getAllTenantIdsSpy.mockResolvedValueOnce(tenantIds);
+      await runner.__._multiTenancyDb();
       await promise;
       expect(acquireLockSpy.mock.calls.filter(([, key]) => key === "EVENT_QUEUE_RUN_PERIODIC_EVENT")).toHaveLength(3);
 
       // second run
       getAllTenantIdsSpy.mockResolvedValueOnce(tenantIds);
-      const p2 = runner.__._multiTenancyDb();
-      await p2.then((p) => Promise.allSettled(p));
+      await runner.__._multiTenancyDb();
       await promisify(setTimeout)(500);
 
       expect(acquireLockSpy.mock.calls.filter(([, key]) => key === "EVENT_QUEUE_RUN_PERIODIC_EVENT")).toHaveLength(3);
@@ -354,16 +346,13 @@ describe("redisRunner", () => {
           }
         });
       });
-      getAllTenantIdsSpy.mockResolvedValueOnce(tenantIds).mockResolvedValueOnce(tenantIds);
-      const p1 = runner.__._multiTenancyDb();
-      await p1.then((p) => Promise.allSettled(p));
+      getAllTenantIdsSpy.mockResolvedValueOnce(tenantIds);
+      await runner.__._multiTenancyDb();
       await promise;
       expect(acquireLockSpy.mock.calls.filter(([, key]) => key === "EVENT_QUEUE_RUN_PERIODIC_EVENT")).toHaveLength(3);
 
       // second run with changed tenant ids
-      getAllTenantIdsSpy
-        .mockResolvedValueOnce(tenantIds.concat("e9bb8ec0-c85e-4035-b7cf-1b11ba8e5792"))
-        .mockResolvedValueOnce(tenantIds.concat("e9bb8ec0-c85e-4035-b7cf-1b11ba8e5792"));
+      getAllTenantIdsSpy.mockResolvedValueOnce(tenantIds.concat("e9bb8ec0-c85e-4035-b7cf-1b11ba8e5792"));
 
       const promise2 = new Promise((resolve) => {
         counter = 0;
@@ -378,8 +367,7 @@ describe("redisRunner", () => {
         });
       });
 
-      const p2 = runner.__._multiTenancyDb();
-      await p2.then((p) => Promise.allSettled(p));
+      await runner.__._multiTenancyDb();
       await promise2;
 
       expect(acquireLockSpy.mock.calls.filter(([, key]) => key === "EVENT_QUEUE_RUN_PERIODIC_EVENT")).toHaveLength(4);
@@ -399,16 +387,13 @@ describe("redisRunner", () => {
           }
         });
       });
-      getAllTenantIdsSpy.mockResolvedValueOnce(tenantIds).mockResolvedValueOnce(tenantIds);
-      const p1 = runner.__._multiTenancyDb();
-      await p1.then((p) => Promise.allSettled(p));
+      getAllTenantIdsSpy.mockResolvedValueOnce(tenantIds);
+      await runner.__._multiTenancyDb();
       await promise;
       expect(acquireLockSpy.mock.calls.filter(([, key]) => key === "EVENT_QUEUE_RUN_PERIODIC_EVENT")).toHaveLength(3);
 
       // second run with changed tenant ids
-      getAllTenantIdsSpy
-        .mockResolvedValueOnce(tenantIds.concat("e9bb8ec0-c85e-4035-b7cf-1b11ba8e5792"))
-        .mockResolvedValueOnce(tenantIds.concat("e9bb8ec0-c85e-4035-b7cf-1b11ba8e5792"));
+      getAllTenantIdsSpy.mockResolvedValueOnce(tenantIds.concat("e9bb8ec0-c85e-4035-b7cf-1b11ba8e5792"));
 
       const promise2 = new Promise((resolve) => {
         counter = 0;
@@ -423,20 +408,16 @@ describe("redisRunner", () => {
         });
       });
 
-      const p2 = runner.__._multiTenancyDb();
-      await p2.then((p) => Promise.allSettled(p));
+      await runner.__._multiTenancyDb();
       await promise2;
 
       expect(acquireLockSpy.mock.calls.filter(([, key]) => key === "EVENT_QUEUE_RUN_PERIODIC_EVENT")).toHaveLength(4);
       acquireLockSpy.mockReset();
 
       // thirds run with same tenant ids
-      getAllTenantIdsSpy
-        .mockResolvedValueOnce(tenantIds.concat("e9bb8ec0-c85e-4035-b7cf-1b11ba8e5792"))
-        .mockResolvedValueOnce(tenantIds.concat("e9bb8ec0-c85e-4035-b7cf-1b11ba8e5792"));
+      getAllTenantIdsSpy.mockResolvedValueOnce(tenantIds.concat("e9bb8ec0-c85e-4035-b7cf-1b11ba8e5792"));
 
-      const p3 = runner.__._multiTenancyDb();
-      await p3.then((p) => Promise.allSettled(p));
+      await runner.__._multiTenancyDb();
       await promisify(setTimeout)(500);
 
       expect(acquireLockSpy.mock.calls.filter(([, key]) => key === "EVENT_QUEUE_RUN_PERIODIC_EVENT")).toHaveLength(0);